chore : 클래스 책임 전이, 테스트 분리
