+++ src/main/java/lotto/domain/Lottos.java
@@ -3,7 +3,6 @@ package lotto.domain;
 import lotto.exception.PrizeOverFlowIsNegativeException;
 
 import java.util.List;
-import java.util.Stack;
 
 public class Lottos {
     private final List<Lotto> lottos;
@@ -29,7 +28,7 @@ public class Lottos {
         int totalPrize = 0;
         int eachPrize = 0;
         for (Lotto lotto : lottos) {
-            eachPrize = lottoWinningStatus(lotto, winningLotto);
+            eachPrize = lotto.lottoWinningStatus(lotto, winningLotto);
             validOverFlow(eachPrize, totalPrize);
             validtotalPrize(eachPrize, totalPrize);
             totalPrize += eachPrize;
@@ -49,38 +48,8 @@ public class Lottos {
         }
     }
 
-
-    private int lottoWinningStatus(Lotto lotto, Lotto winningLotto) {
-        int count = 0;
-        List<LottoNumber> lottoNumbers = lotto.getLotto();
-        List<LottoNumber> winningNumbers = winningLotto.getLotto();
-
-        for (int i = 0; i < winningNumbers.size(); i++) {
-            count += isNumberMatched(winningNumbers, i, lottoNumbers);
-        }
-
-        return Prize.getValueByHit(count);
-    }
-
-    private int isNumberMatched(List<LottoNumber> winningNumbers, int i, List<LottoNumber> lottoNumbers) {
-        if (winningNumbers.contains(lottoNumbers.get(i))) {
-            return 1;
-        }
-        return 0;
-    }
-
     public void additionalLotto(Lotto lotto){
         lottos.add(lotto);
-    }
-
-    public Stack<Integer> calculateStatistic(int totalPrize, Stack<Integer> stack) {
-        int prize = totalPrize;
-        Stack<Integer> statistic = stack;
-        for (Prize price : Prize.values()) {
-            statistic.push( prize / price.getValue());
-            prize %= price.getValue();
-        }
-        return statistic;
     }
 
     public double calculateProfit(int prize, int buyPrice) {
